foaf:account set to trailing /
as
http://orcid.org/0000-0001-6938-0820/orcid-profile
is not accessible, but http://orcid.org/0000-0001-6938-0820/ IS
and returns the profile for browsers.

Thus we can still distinguish between the person and the profile
without rendering two different resources.

Technically
<http://orcid.org/0000-0001-6938-0820> should declare a
HTTP header:

    Content-Location: <http://orcid.org/0000-0001-6938-0820/>

to be crystal clear about the fact that the page differs from the
person, but that is not as important as keeping them separate
within the FOAF rendering in RDF.