Fixes as required by product for typeahead
stopword is not working, but that's OK. We actually want to remove it from the acceptance criteria anyway
LEI is still being indexed, let's remove from index AND filter on query
Not all RINGGOLDS linked to RORs, so lets make that happen
Switch org-disambiguated-name-string and org-disambiguated-name. The string version should have the bigger boost for exact matches.