reverted pom to the release 2.13.16
Jersey library we use is not compatible at runtime with jdk11 config in pom. In order to successfully do the jackson+ swagger lib upgrade and set the maven-compiler-plugin configuration to jdk11, there is the need to upgrade jersey to jersey 2 (https://eclipse-ee4j.github.io/jersey.github.io/documentation/latest/migration.html#mig-1.x ) most probably we need to rework some of the code will not work out of the box upgrading to jersey 2. I will create a separate card for the upgrade.